+++ packages/router/src/bindings/messaging/metaTxRequest.ts
@@ -1,16 +1,15 @@
 import {
   createLoggingContext,
-  MetaTxFulfillPayload,
   MetaTxPayload,
-  MetaTxRouterContractCancelPayload,
-  MetaTxRouterContractFulfillPayload,
-  MetaTxRouterContractPreparePayload,
+  MetaTxTypes,
   MetaTxType,
   NxtpErrorJson,
   RequestContext,
 } from "@connext/nxtp-utils";
 
 import { NoTransactionId } from "../../lib/errors";
+import { TransactionReasons } from "../../lib/entities";
+import { incrementFees, incrementGasConsumed } from "../../lib/helpers";
 import { getOperations } from "../../lib/operations";
 import { getContext } from "../../router";
 
@@ -48,14 +47,9 @@ export const metaTxRequestBinding = async (
     return;
   }
 
-  const transactionId =
-    (data.data as MetaTxFulfillPayload).txData?.transactionId ??
-    (
-      data.data as
-        | MetaTxRouterContractPreparePayload
-        | MetaTxRouterContractFulfillPayload
-        | MetaTxRouterContractCancelPayload
-    ).params?.txData?.transactionId;
+  const { txData, relayerFee } = data.type === MetaTxTypes.Fulfill ? data.data : data.data.params;
+
+  const transactionId = txData.transactionId;
 
   if (!transactionId) {
     throw new NoTransactionId({ data });
@@ -73,6 +67,9 @@ export const metaTxRequestBinding = async (
     const tx = await sendMetaTx(data, requestContext as any);
     if (tx) {
       await messaging.publishMetaTxResponse(from, inbox, { chainId, transactionHash: tx.transactionHash });
+      // Increment collected fees + gas used on relayer fee
+      incrementFees(txData.receivingAssetId, txData.receivingChainId, relayerFee, requestContext);
+      incrementGasConsumed(txData.receivingChainId, tx.gasUsed, TransactionReasons.Relay, requestContext);
     }
     logger.info("Handled fulfill request", requestContext, methodContext);
   } finally {