+++ packages/router/src/lib/entities/index.ts
@@ -16,3 +16,4 @@ export { FulfillInput, FulfillInputSchema } from "./fulfill";
 export { CancelInput, CancelInputSchema } from "./cancel";
 
 export { MetaTxInput, MetaTxInputSchema } from "./metaTx";
+export * from "./metrics";