+++ packages/router/src/lib/helpers/index.ts
@@ -15,3 +15,5 @@ export * from "./shared";
 export { signRouterFulfillTransactionPayload } from "./fulfill";
 
 export { signRouterCancelTransactionPayload } from "./cancel";
+
+export { incrementFees, incrementGasConsumed } from "./metrics";