+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Next Release
 
+- [router] Cleanup prometheus collection
+
+# 0.0.99
+
+- [txservice] Fix error parsing
+
 ## 0.0.98
 
 - [contracts] Set up Moonriver price oracle